turned semantic search into a flag
