+++ yt_fts/yt_fts.py
@@ -1,4 +1,4 @@
-import click, requests, datetime
+import click, requests
 
 from yt_fts.search import get_text, get_text_by_video_id
 from yt_fts.db_utils import * 
@@ -9,7 +9,7 @@ from yt_fts.show import list_channels
 from yt_fts.config import get_config_path, make_config_dir, get_db_path
 
 
-YT_FTS_VERSION = "0.1.25"
+YT_FTS_VERSION = "0.1.26"
 
 @click.group()
 @click.version_option(YT_FTS_VERSION, message='yt_fts version: %(version)s')
@@ -102,46 +102,59 @@ def update(channel, language, number_of_jobs):
         Search for a specified text within a channel, a specific video, or across all channels.
         """
 )
-@click.argument("search_text", required=True)
+@click.argument("text", required=True)
 @click.option("-c", "--channel", default=None, help="The name or id of the channel to search in. This is required unless the --all or --video options are used.")
 @click.option("-v", "--video", default=None, help="The id of the video to search in. This is used instead of the channel option.")
 @click.option("-a", "--all", is_flag=True, help="Search in all channels.")
+@click.option("-s", "--semantic", is_flag=True, help="Use Semantic Search")
+@click.option("-l", "--limit", default=5, help="Max number of results to return")
 @click.option("-e", "--export", is_flag=True, help="Export search results to a CSV file.")
-def search(search_text, channel, video, all, export):
+def search(text, channel, video, all, semantic, limit, export):
 
-    from yt_fts.export import export_search
+    from yt_fts.export import export_fts, export_semantic 
 
-    if len(search_text) > 40:
+    if len(text) > 40:
         show_message("search_too_long")
         exit()
+
+    if channel:
+        scope = "channel"
+        search_id = get_channel_id_from_input(channel)
+    elif video:
+        scope = "video"
+        search_id = video 
+    elif all:
+        scope = "all"
+        search_id = ""
+    
     
     if export:
-        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-
-        if all:
-            file_name = f"all_{timestamp}.csv"
-            export_search(search_text, file_name, search_id="all", scope="all")
-        elif video:
-            file_name = f"video_{video}_{timestamp}.csv"
-            export_search(search_text, file_name, search_id=video, scope="video")
+        if semantic:
+            export_semantic(text, search_id, scope, limit)
         else:
-            channel_id = get_channel_id_from_input(channel)
-            file_name = f"channel_{channel_id}_{timestamp}.csv"
-            export_search(search_text, file_name, search_id=channel_id ,scope="channel")
+            export_fts(text, search_id, scope)
         exit()
+
+
+    if semantic:
+        from yt_fts.search import semantic_search
+        semantic_search(text, search_id, scope, limit)
+        exit()
+
 
     if all:
         print('Searching in all channels')
-        get_text("all", search_text)
+        get_text("all", text)
     elif video:
         print(f"Searching in video {video}")
-        get_text_by_video_id(video, search_text)
+        get_text_by_video_id(video, text)
     elif channel:
         channel_id = get_channel_id_from_input(channel)
-        get_text(channel_id, search_text)
+        get_text(channel_id, text)
     else:
         print("Error: Either --channel, --video, or --all option must be provided")
         exit()
+
 
 # delete
 @click.command( 
@@ -171,70 +184,6 @@ def delete(channel):
         print("Exiting")
 
 
-# Semantic search
-@click.command(
-        help="""
-        Semantic search for specified text.
-
-        Before running this command, you must generate embeddings for the channel using the generate-embeddings command.
-        This command uses OpenAI's embeddings API to search for specified text.
-        An OpenAI API key must be set as an environment variable OPENAI_API_KEY.
-        """
-)
-@click.argument("search_text", required=True)
-@click.option("--channel", default=None, help="channel name or id to search in")
-@click.option("--all", is_flag=True, help="Search all semantic search enabled channels")
-@click.option("--limit", default=3, help="top n results to return")
-def semantic_search(search_text, channel, all, limit):
-
-    from yt_fts.semantic_serch.embeddings import (
-        get_embedding,  
-        save_search_embedding, 
-        search_semantic_search_hist, 
-        check_ss_enabled,
-        )
-
-    from yt_fts.semantic_serch.search_embeddings import (
-        search_using_embedding
-    )
-
-    # check if search string is in semantic search history
-    search_embedding = search_semantic_search_hist(search_text)
-    if search_embedding != None:
-        print("Using cached results")
-    else:
-        print("Generating embeddings for search string")
-        api_key = get_api_key()
-
-        # get embedding for search string and convert to blob
-        print("getting embedding for search string")
-        search_embedding = get_embedding(api_key, search_text)
-
-        # save embedding  search string
-        print("saving embedding for search string")
-        save_search_embedding(search_text, search_embedding)
-
-    if channel != None:
-
-        channel_id = get_channel_id_from_input(channel)
-
-        # verify that embeddings have been created for the channel
-        if check_ss_enabled(channel_id) == False:
-            print("Error: Semantic search not enabled for channel")
-            exit()
-
-        # search using channel id 
-        search_using_embedding(search_embedding, limit, channel_id)
-
-    if all:
-        if check_ss_enabled() == False:
-            print("Error: Semantic search not enabled for any channels")
-            exit()
-
-        # search all ss enabled channels
-        search_using_embedding(search_embedding, limit)
-
-
 # Generate embeddings
 @click.command( 
     help="""
@@ -247,11 +196,9 @@ def semantic_search(search_text, channel, all, limit):
 @click.option("--open-api-key", default=None, help="OpenAI API key. If not provided, the script will attempt to read it from the OPENAI_API_KEY environment variable.")
 def generate_embeddings(channel, open_api_key):
 
-    from yt_fts.semantic_serch.embeddings import (
-        get_openai_embeddings,
-        check_ss_enabled, 
-        enable_ss
-        )
+    from yt_fts.embeddings import get_openai_embeddings
+    from yt_fts.search import check_ss_enabled, enable_ss
+    
 
     channel_id = get_channel_id_from_input(channel)
 
@@ -311,8 +258,8 @@ def show(transcript, channel, library, config):
         list_channels()
 
 
-commands = [download, update, search, semantic_search, 
-            delete, generate_embeddings, show]
+commands = [download, update, search, delete, 
+            generate_embeddings, show]
 
 for command in commands:
     cli.add_command(command)
